HorizontalListLayout, VerticalListLayout, VerticalListFixedRowLayout: allow up to 5 cache misses when calculating the number of visible items
This improves performance by avoiding the case where the size of virtual items needs to be recalculated from scratch. Previously, that could happen when calling dataProvider.updateAt(0) because they only checked that index of the cache. Note: If item sizes are variable, this may still require an extra layout pass because the first measurement will be a little inaccurate. However, it will still avoid destroying all but one item and then recreating them all again. In the bad case, it could still destroy some item renderers, or even create some extras, but it won't have as drastic an effect on performance.